Enable Warnings and clean them

+++ src/client/src/ClientConfiguration.hpp
@@ -14,7 +14,7 @@ namespace onion::voxel
 	{
 		std::string Username = "DefaultUsername";
 		std::string UUID;
-		int RenderDistance = 8;
+		uint8_t RenderDistance = 8;
 	};
 
 	struct ClientConfiguration
@@ -62,7 +62,7 @@ namespace onion::voxel
 
 			clientData.Username = json.value("Username", "");
 			clientData.UUID = json.value("UUID", "");
-			clientData.RenderDistance = json.value("RenderDistance", 8);
+			clientData.RenderDistance = json.value("RenderDistance", static_cast<uint8_t>(8));
 		}
 
 		void Save(const std::filesystem::path& filePath) const
